requirements_laptopenv.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.6
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
audioread==3.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.33.6
botocore==1.33.6
bpemb==0.3.4
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
conllu==4.5.3
cycler==0.11.0
Cython==3.0.6
dataclasses==0.6
datasets==2.13.2
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
dm-tree==0.1.8
etils==0.9.0
filelock==3.12.2
fire==0.5.0
flair==0.12.2
flatbuffers==1.12
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
gast==0.4.0
gdown==4.4.0
gensim==4.2.0
gin-config==0.5.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.48.2
h5py==3.8.0
httplib2==0.22.0
huggingface-hub==0.16.4
hyperopt==0.2.7
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
Janome==0.5.0
jmespath==1.0.1
joblib==1.3.2
kaggle==1.5.16
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langdetect==1.0.9
libclang==16.0.6
librosa==0.8.1
llvmlite==0.36.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.4.3
more-itertools==9.1.0
mpld3==0.3
multidict==6.0.4
multiprocess==0.70.14
networkx==2.6.3
neural-structured-learning==1.4.0
numba==0.53.0
numpy==1.21.6
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.5
Pillow==9.5.0
platformdirs==4.0.0
pooch==1.8.0
pptree==3.1
promise==2.3
proto-plus==1.22.3
protobuf==3.19.6
psutil==5.9.6
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch_revgrad==0.2.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
s3transfer==0.8.2
safetensors==0.4.1
scikit-learn==1.0.2
scipy==1.7.3
segtok==1.5.11
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
sqlitedict==2.1.0
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-addons==0.19.0
tensorflow-datasets==4.8.2
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.3
tensorflowjs==3.18.0
termcolor==2.3.0
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-model-maker==0.3.4
tflite-support==0.4.1
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.1
transformer-smaller-training-vocab==0.2.4
transformers==4.30.2
typeguard==4.1.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==2.2.3
Wikipedia-API==0.6.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.15.0
